Refactored all interfaces into Limebean.Interfaces namespace
