+++ LimeBean.Tests/DatabaseStorageTests_MsSql.cs
@@ -10,6 +10,8 @@ using System.Linq;
 using System.Text;
 using Xunit;
 
+using LimeBean.Interfaces;
+
 namespace LimeBean.Tests {
 
     [Trait("db", "mssql")]
