+++ LimeBean.Tests/DatabaseStorageTests_SQLite.cs
@@ -5,6 +5,8 @@ using System.Linq;
 using System.Text;
 using Xunit;
 
+using LimeBean.Interfaces;
+
 namespace LimeBean.Tests {
 
     public class DatabaseStorageTests_SQLite : IDisposable {
