+++ LimeBean.Tests/RoundtripChecker.cs
@@ -4,6 +4,8 @@ using System.Linq;
 using System.Text;
 using Xunit;
 
+using LimeBean.Interfaces;
+
 namespace LimeBean.Tests {
 
     class RoundtripChecker {
