+++ LimeBean/BeanCrud.cs
@@ -3,6 +3,8 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
+using LimeBean.Interfaces;
+
 namespace LimeBean {
 
     class BeanCrud : IBeanCrud {
