+++ LimeBean/Interfaces/IValueRelaxations.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 
-namespace LimeBean {
+namespace LimeBean.Interfaces {
 
     interface IValueRelaxations {
         bool TrimStrings { get; set; }
