+++ LimeBean/MariaDbDetails.cs
@@ -7,6 +7,8 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 
+using LimeBean.Interfaces;
+
 namespace LimeBean {
 
     class MariaDbDetails : IDatabaseDetails {
