+++ LimeBean/MsSqlDetails.cs
@@ -8,6 +8,8 @@ using System.Linq;
 using System.Text;
 using System.Text.RegularExpressions;
 
+using LimeBean.Interfaces;
+
 namespace LimeBean {
 
     class MsSqlDetails : IDatabaseDetails {
